fix(build): resolve Chisel, C++, and compiler warnings in clean build
This change addresses multiple compilation and build warnings across our first-party Chisel and C++ code, Verilator-generated C++ code, and third-party dependencies to achieve a clean, warning-free build log.

1. FIRST-PARTY CHISEL & SCALA WARNINGS
- Introduced SafeMuxUpTo1H helper in Library.scala which performs safe ChiselEnum casting using the .safe method.
- Updated decoders in Decode.scala to use SafeMuxUpTo1H with respective Enum companion objects (AluOp, BruOp, MluOp, DvuOp, LsuOp, CsrOp), resolving W001 casting warnings.
- Fixed Spi2TLULV2Test.scala to remove an unused local variable captured capture, resolving a Scala compiler warning.
- Introduced a new Custom port type in SoCChiselConfig.scala to allow exposing typed ports (like ChiselEnums) at the subsystem boundary.
- Updated CoralNPUChiselSubsystem.scala to connect external ports directly if their Scala classes match (bypassing asTypeOf), which avoids Chisel's internal asUInt serialization and resolves W001 Enum casting warnings for dm_req_bits_op and dm_rsp_bits_op ports.

2. FIRST-PARTY C++ WARNINGS
- Changed dma_base type to uintptr_t in fpga/sw/display_renderer.cc to prevent -Wint-to-pointer-cast warnings on 64-bit compiler targets.

3. VERILATOR-GENERATED C++ WARNINGS
- Updated local rules_hdl patch 0009-Suppress-Verilator-C-warnings.patch to globally pass -Wno-bool-operation and -Wno-tautological-compare to the compiler for all Verilated libraries.

4. THIRD-PARTY WARNINGS SUPPRESSION
- Added host C++ compiler flag -Wno-sign-compare to .bazelrc to silence protobuf warnings.
- Added Java compiler flags -nowarn and -Xlint:-options,-path to .bazelrc to silence obsolete options and bad path warnings.
- Enabled scalacopts pass-through in chisel_library macro, and passed -nowarn to asyncqueue target in third_party/rocket_chip/BUILD.bazel to silence external Scala compiler warnings.
- Inverted .bazelrc output_filter to ^//(?!(third_party)(/|:|$)) to cleanly hide all remaining external repository warnings (like rules_scala path warnings) while keeping all first-party warnings visible.

TAG=agy

CONV=630c251a-73b1-4570-b9ed-9f4d13425342

+++ fpga/sw/display_renderer.cc
@@ -178,7 +178,7 @@ void DisplayRenderer::Render(const uint16_t* frame_buffer) {
 }
 
 void DisplayRenderer::WaitDma() {
-  uint32_t dma_base = dma_get_base_addr();
+  uintptr_t dma_base = dma_get_base_addr();
   int timeout = 0;
   while (dma_get_status() & 0x1) {
     uint32_t s = dma_get_status();
